+++ src/worker/types.ts
@@ -1,4 +1,5 @@
 import { Job } from "../types/job";
+import { DebugLogger } from "../utils";
 
 export type JobHandler<T = any> = (job: Job<T>) => Promise<void>;
 
@@ -22,4 +23,9 @@ export interface WorkerOptions {
    * Default timezone for cron scheduling
    */
   defaultTimezone?: string;
+
+  /**
+   * Debug logger instance (passed from Scheduler)
+   */
+  debug?: DebugLogger;
 }